added some useful error messages

+++ server/index.js
@@ -174,8 +174,14 @@ io.on("connection", function (socket) {
   socket.on("binarystream", function (m) {
     if (!feedStream) {
       socket.emit("fatal", "rtmp not set yet.");
-      ffmpeg_process.stdin.end();
-      ffmpeg_process.kill("SIGINT");
+      if (ffmpeg_process) {
+        try {
+          ffmpeg_process.stdin.end();
+          ffmpeg_process.kill("SIGINT");
+        } catch (e) {
+          console.warn("killing ffmoeg process attempt failed...");
+        }
+      }
       return;
     }
     feedStream(m);
